+++ lib/nerves/artifact/provider.ex
@@ -5,14 +5,18 @@ defmodule Nerves.Artifact.Provider do
   A package provider is a module that can take package source and produce
   artifacts.
   """
+  
+  @type build_result :: {:ok, build_path :: String.t} | {:error, reason :: term}
+  @type archive_result :: {:ok, path :: String.t} | {:error, reason :: term}
+  @type clean_result :: :ok | {:error, reason :: term}
 
   @callback build(package :: Nerves.Package.t, toolchain :: atom, opts :: term) ::
-    {:ok, build_path :: String.t} | {:error, reason :: term}
+    build_result
 
   @callback archive(package :: Nerves.Package.t, toolchain :: atom, opts :: term) ::
-    {:ok, path :: String.t} | {:error, reason :: term}
+    archive_result
 
   @callback clean(package :: Nerves.Package.t) ::
-    :ok | {:error, reason :: term}
+    clean_result
 
 end