+++ lib/nerves/env.ex
@@ -300,14 +300,23 @@ defmodule Nerves.Env do
     if nerves_system_path != nil and File.dir?(nerves_system_path) do
       # Bootstrap the build platform
       platform = Nerves.Env.system.platform
-      # Pre 0.4.0 Legacy
-      platform = platform || Nerves.Env.system.config[:build_platform]
-      ## end re 0.4.0 Legacy
+
       pkg =
         Nerves.Env.packages_by_type(:system_platform)
         |> List.first
       platform.bootstrap(pkg)
     end
+
+    # Bootstrap all other packahes who define a platform
+    Nerves.Env.packages
+    |> Enum.reject(& &1 == Nerves.Env.toolchain())
+    |> Enum.reject(& &1 == Nerves.Env.system())
+    |> Enum.reject(& &1.platform == nil)
+    |> Enum.each(fn
+      (%{platform: platform} = pkg) -> 
+        platform.bootstrap(pkg)
+      _ -> :noop
+    end)
   end
 
   @doc false